refactor: rename WATF part 2
