+++ src/compiler/parsing/wast/index.js
