+++ src/compiler/parsing/wast/number-literals.js
