+++ src/compiler/parsing/wast/string-literals.js
