+++ src/compiler/parsing/wast/tokenizer.js
